Explore the National Emissions Inventory database
The goal of this assignment is to explore the National Emissions Inventory database and see what it say about fine particulate matter pollution in the United states over the 10-year period 1999–2008. The most inresting finding is that the Clean Air Act programs have lowered national concentrations of P2.5 from coal by over 2e+05 Tons between 1999 and 2008, and the good news is that this trend has continued.